versioon

+++ main.py
@@ -48,26 +48,36 @@ async def search(message):
     queue_info = db.get_queue(message.from_user.id)
     if chat_info == False:
         if queue_info == False:
-            markup = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
-            button1 = types.KeyboardButton(cfg.STOP_SEARCH)
-            markup.add(button1)
-
-            chat_two = db.get_user_queue()
-
-            if db.create_chat(message.from_user.id, chat_two) == False:
-                db.add_queue(message.from_user.id)
-                await message.answer(cfg.SEARCH_PROCESS, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
-            else:
-                try:
-                    await dp.bot.send_message(message.from_user.id, cfg.SEARCH_TRUE, reply_markup=types.ReplyKeyboardRemove(), parse_mode=types.ParseMode.MARKDOWN)
-                    await dp.bot.send_message(chat_two, cfg.SEARCH_TRUE, reply_markup=types.ReplyKeyboardRemove(), parse_mode=types.ParseMode.MARKDOWN)
-                except BotBlocked:
-                    db.delete_chat(message.from_user.id)
+            chann = db.get_channels()
+            admin_channels = await get_admin_channels(chann)
+            text = cfg.TEXT_SUBCRIBE
+            for i, item in enumerate(admin_channels, 1):
+                text += f"\n {i}. {item}"
+            for i in admin_channels:
+                subscribded = await check_user_subscription(i, message.from_user.id)
+                if subscribded:
                     markup = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
-                    button1 = types.KeyboardButton(cfg.SEARCH)
-                    button2 = types.KeyboardButton(cfg.SEARCH_MALE)
-                    markup.add(button1, button2)
-                    await message.answer(cfg.BOT_BLOCKED, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
+                    button1 = types.KeyboardButton(cfg.STOP_SEARCH)
+                    markup.add(button1)
+
+                    chat_two = db.get_user_queue()
+
+                    if db.create_chat(message.from_user.id, chat_two) == False:
+                        db.add_queue(message.from_user.id)
+                        await message.answer(cfg.SEARCH_PROCESS, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
+                    else:
+                        try:
+                            await dp.bot.send_message(message.from_user.id, cfg.SEARCH_TRUE, reply_markup=types.ReplyKeyboardRemove(), parse_mode=types.ParseMode.MARKDOWN)
+                            await dp.bot.send_message(chat_two, cfg.SEARCH_TRUE, reply_markup=types.ReplyKeyboardRemove(), parse_mode=types.ParseMode.MARKDOWN)
+                        except BotBlocked:
+                            db.delete_chat(message.from_user.id)
+                            markup = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
+                            button1 = types.KeyboardButton(cfg.SEARCH)
+                            button2 = types.KeyboardButton(cfg.SEARCH_MALE)
+                            markup.add(button1, button2)
+                            await message.answer(cfg.BOT_BLOCKED, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
+                else:
+                    await message.answer(text)
         else:
             await message.answer(cfg.CANCEL_SEARCH_PROCESS, parse_mode=types.ParseMode.MARKDOWN)
     else:
@@ -180,24 +190,34 @@ async def next(message: types.Message):
         chat_info = db.get_active_chat(message.from_user.id)
         queue_info = db.get_queue(message.from_user.id)
         if chat_info != False:
-            db.delete_chat(message.from_user.id)
-            markups = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
-            buttons1 = types.KeyboardButton(cfg.SEARCH)
-            markups.add(buttons1)
-            await dp.bot.send_message(chat_info, cfg.STOP_DIALOG_TEXT_SOBESEDNIK, reply_markup=markups, parse_mode=types.ParseMode.MARKDOWN)
+            chann = db.get_channels()
+            admin_channels = await get_admin_channels(chann)
+            text = cfg.TEXT_SUBCRIBE
+            for i, item in enumerate(admin_channels, 1):
+                text += f"\n {i}. {item}"
+            for i in admin_channels:
+                subscribded = await check_user_subscription(i, message.from_user.id)
+                if subscribded:
+                    db.delete_chat(message.from_user.id)
+                    markups = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
+                    buttons1 = types.KeyboardButton(cfg.SEARCH)
+                    markups.add(buttons1)
+                    await dp.bot.send_message(chat_info, cfg.STOP_DIALOG_TEXT_SOBESEDNIK, reply_markup=markups, parse_mode=types.ParseMode.MARKDOWN)
 
-            markup = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
-            button1 = types.KeyboardButton(cfg.STOP_SEARCH)
-            markup.add(button1)
+                    markup = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
+                    button1 = types.KeyboardButton(cfg.STOP_SEARCH)
+                    markup.add(button1)
 
-            chat_two = db.get_user_queue()
+                    chat_two = db.get_user_queue()
 
-            if db.create_chat(message.from_user.id, chat_two) == False:
-                db.add_queue(message.from_user.id)
-                await message.answer(cfg.SEARCH_PROCESS, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
-            else:
-                await dp.bot.send_message(message.from_user.id, cfg.SEARCH_TRUE, parse_mode=types.ParseMode.MARKDOWN)
-                await dp.bot.send_message(chat_two, cfg.SEARCH_TRUE, parse_mode=types.ParseMode.MARKDOWN)
+                    if db.create_chat(message.from_user.id, chat_two) == False:
+                        db.add_queue(message.from_user.id)
+                        await message.answer(cfg.SEARCH_PROCESS, reply_markup=markup, parse_mode=types.ParseMode.MARKDOWN)
+                    else:
+                        await dp.bot.send_message(message.from_user.id, cfg.SEARCH_TRUE, parse_mode=types.ParseMode.MARKDOWN)
+                        await dp.bot.send_message(chat_two, cfg.SEARCH_TRUE, parse_mode=types.ParseMode.MARKDOWN)
+                else:
+                    await message.answer(text)
         elif queue_info != False:
             await message.answer(cfg.CANCEl_STOP_SEARCH_TEXT, parse_mode=types.ParseMode.MARKDOWN)
         else:
